Add GStore & GStoreMut composite store trait,
Split Store into Store + Index and StoreMut into StoreMut + IndexMut.
GStore trait represents Store + Index.
GStoreMut trait represents StoreMut + IndexMut + AlterTable.
Now execution layer codes only use GStore and GStoreMut,
so user does not need to care anymore whether which feature is enabled or not.
