+++ src/storages/sled_storage/mod.rs
@@ -1,14 +1,16 @@
 mod alter_table;
 mod error;
-#[cfg(feature = "index")]
+mod index;
+mod index_mut;
 mod index_sync;
 mod store;
 mod store_mut;
-#[cfg(not(feature = "alter-table"))]
-impl crate::AlterTable for SledStorage {}
 
 use {
-    crate::{Error, Result, RowIter, Schema},
+    crate::{
+        store::{GStore, GStoreMut},
+        Error, Result, RowIter, Schema,
+    },
     error::err_into,
     sled::{self, Config, Db, IVec},
     std::convert::TryFrom,
@@ -60,3 +62,6 @@ fn scan_data(tree: &Db, table_name: &str) -> RowIter<IVec> {
 
     Box::new(result_set)
 }
+
+impl GStore<IVec> for SledStorage {}
+impl GStoreMut<IVec> for SledStorage {}
