+++ src/storages/sled_storage/store_mut.rs
@@ -9,11 +9,7 @@ use {
 };
 
 #[cfg(feature = "index")]
-use {
-    super::{fetch_schema, index_sync::IndexSync, scan_data},
-    crate::{ast::Expr, IndexError, Result},
-    std::iter::once,
-};
+use {super::index_sync::IndexSync, crate::IndexError};
 
 #[cfg(feature = "index")]
 macro_rules! try_self {
@@ -179,113 +175,6 @@ impl StoreMut<IVec> for SledStorage {
                     .map_err(ConflictableTransactionError::Abort)?;
 
                 index_sync.delete(&tree, &key, &row)?;
-            }
-
-            Ok(())
-        })
-    }
-
-    #[cfg(feature = "index")]
-    async fn create_index(
-        self,
-        table_name: &str,
-        index_name: &str,
-        index_expr: &Expr,
-    ) -> MutResult<Self, ()> {
-        let (schema_key, schema) = try_self!(self, fetch_schema(&self.tree, table_name));
-        let Schema {
-            column_defs,
-            indexes,
-            ..
-        } = try_into!(
-            self,
-            schema.ok_or_else(|| IndexError::TableNotFound(table_name.to_owned()))
-        );
-
-        if indexes.iter().any(|(name, _)| name == index_name) {
-            return Err((
-                self,
-                IndexError::IndexNameAlreadyExists(index_name.to_owned()).into(),
-            ));
-        }
-
-        let indexes = indexes
-            .into_iter()
-            .chain(once((index_name.to_owned(), index_expr.clone())))
-            .collect::<Vec<_>>();
-
-        let schema = Schema {
-            table_name: table_name.to_owned(),
-            column_defs,
-            indexes,
-        };
-
-        let rows = try_self!(
-            self,
-            scan_data(&self.tree, table_name).collect::<Result<Vec<_>>>()
-        );
-        let index_sync = IndexSync::from(&schema);
-
-        transaction!(self, |tree| {
-            let schema_value = bincode::serialize(&schema)
-                .map_err(err_into)
-                .map_err(ConflictableTransactionError::Abort)?;
-
-            tree.insert(schema_key.as_bytes(), schema_value)?;
-
-            for (data_key, row) in rows.iter() {
-                index_sync.insert(&tree, data_key, row)?;
-            }
-
-            Ok(())
-        })
-    }
-
-    #[cfg(feature = "index")]
-    async fn drop_index(self, table_name: &str, index_name: &str) -> MutResult<Self, ()> {
-        let (schema_key, schema) = try_self!(self, fetch_schema(&self.tree, table_name));
-        let Schema {
-            column_defs,
-            indexes,
-            ..
-        } = try_into!(
-            self,
-            schema.ok_or_else(|| IndexError::TableNotFound(table_name.to_owned()))
-        );
-
-        if indexes.iter().all(|(name, _)| name != index_name) {
-            return Err((
-                self,
-                IndexError::IndexNameDoesNotExist(index_name.to_owned()).into(),
-            ));
-        }
-
-        let indexes = indexes
-            .into_iter()
-            .filter(|(name, _)| name != index_name)
-            .collect::<Vec<_>>();
-
-        let schema = Schema {
-            table_name: table_name.to_owned(),
-            column_defs,
-            indexes,
-        };
-
-        let rows = try_self!(
-            self,
-            scan_data(&self.tree, table_name).collect::<Result<Vec<_>>>()
-        );
-        let index_sync = IndexSync::from(&schema);
-
-        transaction!(self, |tree| {
-            let schema_value = bincode::serialize(&schema)
-                .map_err(err_into)
-                .map_err(ConflictableTransactionError::Abort)?;
-
-            tree.insert(schema_key.as_bytes(), schema_value)?;
-
-            for (data_key, row) in rows.iter() {
-                index_sync.delete(&tree, data_key, row)?;
             }
 
             Ok(())
