+++ src/store/mod.rs
@@ -1,14 +1,38 @@
-#[cfg(feature = "alter-table")]
-mod alter_table;
-#[cfg(feature = "alter-table")]
-pub use alter_table::*;
-#[cfg(not(feature = "alter-table"))]
-pub trait AlterTable {}
+use cfg_if::cfg_if;
 
-#[cfg(feature = "index")]
-mod index;
-#[cfg(feature = "index")]
-pub use index::IndexError;
+cfg_if! {
+    if #[cfg(feature = "alter-table")] {
+        mod alter_table;
+        pub use alter_table::{AlterTable, AlterTableError};
+    }
+}
+
+cfg_if! {
+    if #[cfg(feature = "index")] {
+        mod index;
+        pub use index::{Index, IndexError, IndexMut};
+    }
+}
+
+cfg_if! {
+    if #[cfg(feature = "index")] {
+        pub trait GStore<T: Debug>: Store<T> + Index<T> {}
+    } else {
+        pub trait GStore<T: Debug>: Store<T> {}
+    }
+}
+
+cfg_if! {
+    if #[cfg(all(feature = "alter-table", feature = "index"))] {
+        pub trait GStoreMut<T: Debug>: StoreMut<T> + IndexMut<T> + AlterTable {}
+    } else if #[cfg(feature = "alter-table")] {
+        pub trait GStoreMut<T: Debug>: StoreMut<T> + AlterTable {}
+    } else if #[cfg(feature = "index")] {
+        pub trait GStoreMut<T: Debug>: StoreMut<T> + IndexMut<T> {}
+    } else {
+        pub trait GStoreMut<T: Debug>: Store<T> + StoreMut<T> {}
+    }
+}
 
 use {
     crate::{
@@ -19,33 +43,18 @@ use {
     std::fmt::Debug,
 };
 
-#[cfg(feature = "index")]
-use crate::{
-    ast::{Expr, IndexOperator},
-    data::Value,
-};
-
 pub type RowIter<T> = Box<dyn Iterator<Item = Result<(T, Row)>>>;
 
-/// By implementing `Store` trait, you can run `SELECT` queries.
+/// By implementing `Store` trait, you can run `SELECT` query.
 #[async_trait(?Send)]
 pub trait Store<T: Debug> {
     async fn fetch_schema(&self, table_name: &str) -> Result<Option<Schema>>;
 
     async fn scan_data(&self, table_name: &str) -> Result<RowIter<T>>;
-
-    #[cfg(feature = "index")]
-    async fn scan_indexed_data(
-        &self,
-        table_name: &str,
-        index_name: &str,
-        op: &IndexOperator,
-        value: Value,
-    ) -> Result<RowIter<T>>;
 }
 
-/// `StoreMut` takes role of mutation, related to `INSERT`, `CREATE`, `DELETE`, `DROP` and
-/// `UPDATE`.
+/// By implementing `StoreMut` trait,
+/// you can run `INSERT`, `CREATE TABLE`, `DELETE`, `UPDATE` and `DROP TABLE` queries.
 #[async_trait(?Send)]
 pub trait StoreMut<T: Debug>
 where
@@ -60,15 +69,4 @@ where
     async fn update_data(self, table_name: &str, rows: Vec<(T, Row)>) -> MutResult<Self, ()>;
 
     async fn delete_data(self, table_name: &str, keys: Vec<T>) -> MutResult<Self, ()>;
-
-    #[cfg(feature = "index")]
-    async fn create_index(
-        self,
-        table_name: &str,
-        index_name: &str,
-        column: &Expr,
-    ) -> MutResult<Self, ()>;
-
-    #[cfg(feature = "index")]
-    async fn drop_index(self, table_name: &str, index_name: &str) -> MutResult<Self, ()>;
 }
